Move global Psbt fields to a Global type
The PSBT conceptually has three pieces

1. The global map
2. The input map.
3. The output map.

We currently have `Input` and `Output` for the input/output
maps (vectors actually) but we inline all the global map pairs into the
`Psbt` struct.

As a step towards simplifying the code add a `Global` type that
abstracts away all the stuff for the glabal map.