Re-import bitcoin and miniscript code
We would like the PSBTv0 code to be as close as possible to the code
we originally imported from `rust-bitcoin` and `rust-miniscript`. This
gives us a few things:

- Same code means downstream can easily switch from using the `psbt`
  modules in `rust-bitcoin` and `rust-miniscritp`, hopefully just need
  to change import paths.
- Re-basing on new versions of the original code will be easier and
  less bug-prone if the files stay as close to the same as possible.

Re-import the original code, make as few a changes as possible to get
it to build. Use the same directory structure as the original, unlike
what we did when we originally imported the code.

Note, this patch is not PSBTv2 compliant because deserialization of v0
PSBTs does not fail for excluded map keys. That will be added in a
subsequent patch.