Upgrade to latest bitcoin
Includes upgrade to:

- bitcoin v0.32.2
- miniscript v12.2.0
- secp256k1 v0.29.0
- bitcoin-internals v0.3.0

To upgrade we need to make the v0 module mirror the code in
rust-bitcoin::psbt, by diffing the files that is reasonably easy to do
and only slightly error prone - would be great if someone else would
diff and view as well.

It gets wild in the v2 module because we copied code from the v0
module then modified it for PSBT v2. All changes to the v2 module need
going over very carefully.